pyproject.toml
[tool.poetry]
name = "report-hub"
version = "0.1.0"
description = "ReportHub by IMMAP"
authors = ["RH-Team <rh-support@immap.org>"]
license = "GPL"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0.6"
django-smart-selects = {git = "https://github.com/jazzband/django-smart-selects.git"}
django-filter = "24.2"
Faker = "14.2.1"
python-dateutil = "2.8.2"
six = "1.16.0"
pandas = "1.4.3"
django-environ = "0.10.0"
django-compressor = "4.3.1"
bcrypt = "4.0.1"
openpyxl = "3.1.2"
pytest = "^7.4.0"
flake8 = "^6.0.0"
2to3 = "1.0"
asgiref = "3.8.1"
astroid = "2.12.10"
attrs = "22.1.0"
certifi = "2022.6.15.1"
charset-normalizer = "2.1.1"
dill = "0.3.5.1"
django-appconf = "1.0.5"
django-cors-headers = "3.13.0"
django-jsonform = "2.12.0"
django-jsonforms = "1.1.2"
djangorestframework = "3.13.1"
greenlet = "2.0.2"
idna = "3.3"
isort = "5.10.1"
jsonschema = "4.16.0"
lazy-object-proxy = "1.7.1"
mccabe = "0.7.0"
model-bakery = "1.7.0"
mypy-extensions = "0.4.3"
numpy = "1.23.1"
pycparser = "2.21"
pylint = "2.15.3"
pylint-django = "2.5.3"
pylint-plugin-utils = "0.7"
pyrsistent = "0.18.1"
pytz = "2022.1"
PyYAML = "6.0"
rcssmin = "1.1.1"
requests = "2.28.1"
rjsmin = "1.2.1"
sqlparse = "0.4.2"
tomli = "2.0.1"
tomlkit = "0.11.4"
typing_extensions = "4.7.0"
urllib3 = "1.26.12"
wrapt = "1.14.1"
django-vite-plugin = "^1.3.3"
platformdirs = "^3.11.0"
virtualenv = "^20.24.6"
django-import-export = "^3.3.4"
factory-boy = "^3.3.0"
sentry-sdk = {extras = ["django"], version = "^1.39.1"}
ruff = "^0.2.1"
django-htmx = "^1.17.2"
psycopg2-binary = "^2.9.9"
django-guardian = "^2.4.0"
django-dbbackup = "^4.1.0"
dropbox = "^12.0.0"
django-storages = "^1.14.3"
django-extra-settings = "^0.12.0"


[tool.poetry.group.dev.dependencies]
django-extensions = "^3.2.3"
nplusone = "^1.0.0"
djlint = "^1.34.0"
django-tui = "^23.9"
django-debug-toolbar = "^4.2.0"


[tool.ruff]
# Exclude a variety of commonly ignored directories.
line-length=120
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "scripts",
    "migrations"
]



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
